# api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/api/gears": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gear"
                ],
                "summary": "List Gears",
                "parameters": [
                    {
                        "type": "string",
                        "description": "pagination limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "active page",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        " count": {
                                            "type": "integer"
                                        },
                                        "Results": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Gear"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/gears/{id}": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "gear"
                ],
                "summary": "Get Gear",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Gear ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Gear"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.PaginatedResponse": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "results": {}
            }
        },
...
        "models.Athlete": {
            "type": "object",
            "properties": {
                "approve_followers": {
                    "description": "if has enhanced privacy enabled",
                    "type": "boolean"
                },
                "badge_type_id": {
                    "type": "integer"
                },
                "city": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "firstname": {
                    "type": "string"
                },
                "follower": {
                    "description": "this athlete’s following status of the authenticated athlete",
                    "type": "string"
                },
                "friend": {
                    "description": "‘pending’, ‘accepted’, ‘blocked’ or ‘null’, the authenticated athlete’s following status of this athlete",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastname": {
                    "type": "string"
                },
                "premium": {
                    "type": "boolean"
                },
                "profile": {
                    "description": "URL to a 124x124 pixel profile picture",
                    "type": "string"
                },
                "profile_medium": {
                    "description": "URL to a 62x62 pixel profile picture",
                    "type": "string"
                },
                "sex": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "tx": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.Gear": {
            "type": "object",
            "properties": {
                "athlete": {
                    "$ref": "#/definitions/models.Athlete"
                },
                "athleteID": {
                    "type": "integer"
                },
                "brand_name": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "distance": {
                    "type": "number"
                },
                "frame_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "model_name": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "primary": {
                    "type": "boolean"
                }
            }
        },
...
    }
}